add week parity

+++ balder_schedule_app/lib/widgets/edit/lesson/lesson_field.dart
@@ -29,7 +29,7 @@ class LessonField extends StatelessWidget {
           return 'Поле не может быть пустым';
         }
 
-        if (isNumeric && double.tryParse(value) == null) {
+        if (isNumeric && int.tryParse(value) == null) {
           return 'Введите число';
         }
